feature: moving location service to service select
