+++ libs/api/controllers/src/index.ts
@@ -1,5 +1,4 @@
 export * from './auth';
 export * from './student';
 export * from './user';
-export * from './location';
 export * from './select';
