+++ libs/api/master/src/index.ts
@@ -2,7 +2,6 @@ import { Module } from '@nestjs/common';
 import { PrismaModule } from '@uninus/api/models';
 import {
   AuthModule,
-  LocationModule,
   PmbModule,
   SelectModule,
   StudentModule,
@@ -16,7 +15,6 @@ import {
     PmbModule,
     StudentModule,
     UserModule,
-    LocationModule,
     SelectModule,
   ],
   controllers: [],
