+++ libs/api/services/src/index.ts
@@ -5,6 +5,5 @@ export * from './email';
 export * from './student';
 export * from './user';
 export * from './cloudinary';
-export * from './location';
 export * from './swagger';
 export * from './select';
