+++ libs/entities/src/types/select-field.ts
@@ -2,6 +2,10 @@ export interface ISelectRequest {
   search: string;
 }
 
+export interface IProvinceRequest {
+  search: string;
+}
+
 export interface ISelectFacultyRequest extends ISelectRequest {
   degree_program_id: string;
 }
@@ -9,6 +13,35 @@ export interface ISelectFacultyRequest extends ISelectRequest {
 export interface ISelectDepartmentRequest extends ISelectRequest {
   faculty_id: string;
 }
+
+export type TProvinceResponse = {
+  province: Array<{
+    id: number;
+    name: string;
+  }>;
+};
+
+export interface ICityRequest extends ISelectRequest {
+  province_id: string;
+}
+
+export type TCityResponse = {
+  city: Array<{
+    id: number;
+    name: string;
+  }>;
+};
+
+export interface ISubDistrictRequest extends ISelectRequest {
+  city_id: string;
+}
+
+export type TSubDistrictResponse = {
+  sub_district: Array<{
+    id: number;
+    name: string;
+  }>;
+};
 
 export type TDegreeProgramResponse = {
   degree_program: Array<{
